+++ bootique-job/src/main/java/io/bootique/job/JobsFactory.java
@@ -97,7 +97,7 @@ public class JobsFactory {
     private Map<String, Object> createDefaultParams(JobMetadata metadata) {
 
         if (metadata.getParameters().isEmpty()) {
-            return Collections.emptyMap();
+            return Map.of();
         }
 
         Map<String, Object> map = new HashMap<>();